tests

+++ FinanceManager.PL.MVC/Models/AccountViewModel.cs
@@ -4,6 +4,6 @@ namespace FinanceManager.PL.MVC.Models
     {
         public string Number { get; set; }
         
-        public int Count { get; set; }
+        public decimal Count { get; set; }
     }
 }
